# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.StatusPayload:
    properties:
      comment:
        type: string
      status:
        type: string
    type: object
  dto.TripPayload:
    properties:
      comment:
        type: string
      currency:
        type: string
      destination_city:
        type: string
      destination_country:
        type: string
      end_date:
        type: string
      planned_daily_allowance:
        type: number
      planned_hotel:
        type: number
      planned_other:
        type: number
      planned_transport:
        type: number
      project_id:
        type: string
      purpose:
        type: string
      start_date:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for the Business Trip and Expense Portal
  title: Travel Portal API
  version: "1.0"
paths:
...
  /trip-requests:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List trip requests
      tags:
      - Trips
    post:
      consumes:
      - application/json
      parameters:
      - description: Trip data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.TripPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create trip request
      tags:
      - Trips
  /trip-requests/{id}:
    delete:
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete trip
      tags:
      - Trips
    get:
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get trip by ID
      tags:
      - Trips
    put:
      consumes:
      - application/json
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: string
      - description: Trip data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.TripPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update trip
      tags:
      - Trips
  /trip-requests/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: string
      - description: Status payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.StatusPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change trip status
      tags:
      - Trips
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"